Make fetch_wordlist.sh more user-friendly

+++ scripts/fetch_wordlist.sh
@@ -11,19 +11,24 @@ set -e
 usage() {
   cat <<USAGE
   Usage: $0 [options]
-     Fetch dictionary files for specified language and convert to a wordlist
+    Fetch dictionary files for specified language and convert to a wordlist
   Options:
     -h, --help:
           print this message
     -l, --lang:
           (required) language to generate wordlist for
+          options: ar, en, es, fr, hi, pt, ru, sw
     -d, --dry-run:
-          print the commands instead of executing them
+          print commands instead of executing them
     -v, --verbose:
-          print extra internal variable values to help with debugging
-
+          print each line of code before it is executed
   Caveats:
-    This script assumes internet access and hunspell reader has been installed with npm
+    This script assumes:
+      * internet access
+      * scripts.wordlist defined in package.json
+      * hunspell-reader installed with npm
+    If you run this script many times in rapid succession,
+      your dictionary download may be throttled by the source
 USAGE
 }
 
@@ -64,7 +69,9 @@ SRC=https://cgit.freedesktop.org/libreoffice/dictionaries/plain/
 case ${LANG} in
   ar)
     # URL=${SRC}ar/ar
-    echo "Manually download the wordlist from https://sourceforge.net/projects/arabic-wordlist/"
+    echo "The Arabic LibreOffice dictionary generates a 1.5 GB wordlist."
+    echo "Manually download the wordlist from https://sourceforge.net/projects/arabic-wordlist/ instead"
+    echo "and save it to src/resources/dictionaries/ar.txt for use with the split_dictionary.py script."
     exit 1
     ;;
   en)
@@ -90,13 +97,14 @@ case ${LANG} in
     ;;
   *)
     echo "Unavailable language: ${LANG}"
+    echo "Options: ar, en, es, fr, hi, pt, ru, sw"
     exit 1
     ;;
 esac
 
 DIR=src/resources/dictionaries/
 
-echo "Fetching .aff file"
+echo "** Fetching .aff file **"
 AFF=${DIR}${LANG}.aff
 cmd="curl -o ${AFF} ${URL}.aff"
 if [[ $DRYRUN -eq 1 ]] ; then
@@ -105,7 +113,7 @@ else
   $cmd
 fi
 
-echo "Fetching .dic file"
+echo "** Fetching .dic file **"
 DIC=${DIR}${LANG}.dic
 cmd="curl -o ${DIC} ${URL}.dic"
 if [[ $DRYRUN -eq 1 ]] ; then
@@ -114,13 +122,21 @@ else
   $cmd
 fi
 
-echo "Converting to .txt wordlist"
-cmd="npm run wordlist -- ${DIC} -po ${DIR}${LANG}.txt"
+echo "** Converting to .txt wordlist **"
+TXT=${DIR}${LANG}.txt
+cmd="npm run wordlist -- ${DIC} -po ${TXT}"
 if [[ $DRYRUN -eq 1 ]] ; then
   echo "$cmd"
 else
   $cmd
 fi
+echo "** Wordlist saved to ${TXT} **"
 
-rm $AFF
-rm $DIC
+if [[ $DRYRUN -eq 1 ]] ; then
+  echo "rm $AFF"
+  echo "rm $DIC"
+else
+  rm $AFF
+  rm $DIC
+fi
+echo "** Deleted .aff and .dic files **"
